backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==3.2.25
django-extensions==3.2.3
django-stubs==5.1.0
django-stubs-ext==5.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.3
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
flake8-bugbear==24.8.19
flake8-builtins==2.5.0
flake8-comprehensions==3.15.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-quotes==3.4.0
flake8-return==1.2.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.14.1
pillow==10.4.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
